backend/requirements.txt
blinker==1.9.0
click==8.1.8
defusedxml==0.7.1
fastexcel==0.15.1
Flask==3.1.0
flask-cors==6.0.1
fonttools==4.60.0
fpdf2==2.8.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
polars==1.33.0
psycopg==3.2.6
psycopg-binary==3.2.6
pyarrow==21.0.0
python-barcode==0.16.1
python-dotenv==1.1.0
typing_extensions==4.13.2
watchdog==6.0.0
Werkzeug==3.1.3
